wspace/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
markdown==3.5.1
google-api-python-client==2.111.0
google-auth-oauthlib==1.2.0
